CORECCIONES V

+++ server/src/services/empleado.services.js
@@ -1,4 +1,5 @@
 import { Usuario } from "../models/Usuario.js";
+import { Venta } from "../models/Venta.js";
 import { Empleado } from "../models/Empleado.js";
 import { EmpleadoDTO } from "../dtos/empleado.dto.js";
 import bcrypt from "bcryptjs";
@@ -134,6 +135,16 @@ export async function actualizarEmpleado(
 //Eliminar Empleado de un usuario
 export async function eliminarEmpleado(idEmpleado, idUsuario) {
   try {
+    const ventas = await Venta.findAll({
+      where: {
+        idEmpleado: idEmpleado,
+      },
+    });
+
+    if (ventas.length > 0) {
+      throw new Error("No se puede eliminar el empleado, tiene ventas asociadas");
+    }
+
     await Empleado.destroy({
       where: {
         idEmpleado: idEmpleado,
